+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Separate clean up functions to their own file (T-13957)
 * Update PUC to v5.5 (T-20922)
 * Prevent add media files from dev if there is a staging environment or files have been added from staging (T-15169)
+* Update media URL and path staging and dev environments properly (T-15169)
 
 ### 3.1.1: 2024-12-02